% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_workout.R
\name{generate_workout}
\alias{generate_workout}
\title{Generate workout}
\usage{
generate_workout(n_workout, complexity = 1, workout = wk, inside_only = TRUE)
}
\arguments{
\item{n_workout}{Total number of workout (2 cycles)}

\item{complexity}{The complexity of the exercises}

\item{workout}{Workout database object}

\item{inside_only}{Keep only inside friendly activities}
}
\value{

}
\description{
Generate workout
}
