% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_workout.R
\name{post_workout}
\alias{post_workout}
\title{Post Workout as a GitHub issue}
\usage{
post_workout(
  owner,
  repo,
  body = generate_workout(16),
  title = sprintf("Workout - \%s", Sys.Date()),
  assignee = owner,
  ...
)
}
\arguments{
\item{owner}{Owner of the repo}

\item{repo}{Name of the repo}

\item{body}{Content of the Workout}

\item{title}{Title of the issue}

\item{assignee}{Issue assignee}

\item{...}{other arguments that can be passed to `gh::gh()`}
}
\value{

}
\description{
Post Workout as a GitHub issue
}
